Support /models endpoint
This change handles the /models endpoint.
The old site had a /models endoint that handled
search categories for most of the category types.
With the turnover 404 counts I can see that we have
users for that page still.

I added a new template that closely resembles the
old template but is better because it supports
options for all of the categories on the navbar.