Support lowercase in make names
The old site checked for any case in make names.
It didn't handle makes with an spaces and other
characters.

For the future, some redesign should happen so
that we support normalized names. We can't use
the vendor query param for this since that is
too permissive and will return, e.g. Dell and
Dell EMC.

The API should be changed to handle normalized
makes.